requirements.txt
anyio==4.13.0
certifi==2026.4.22
charset-normalizer==3.4.7
colorama==0.4.6
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.15
numpy==2.4.5
python-telegram-bot==22.7
requests==2.34.2
typing_extensions==4.15.0
urllib3==2.7.0
